requirements.txt
streamlit
numpy
pandas
scikit-learn
joblib
matplotlib
plotly
seaborn
xgboost
PyPDF2
fpdf2
